Revert "Provide more agent analytics to posthog (#10887)"
This reverts commit a1ce482f3e336dbabe1b12b92f6339af6a2bbbf8.
