+++ daemon/common.h
@@ -82,7 +82,6 @@
 #include "main.h"
 #include "signals.h"
 #include "commands.h"
-#include "analytics.h"
 
 // global netdata daemon variables
 extern char *netdata_configured_hostname;
